update uploaded file deletion and minor redirect url adjustment

+++ application/models/AnalysisModel.php
@@ -36,12 +36,13 @@ class AnalysisModel extends CI_Model
         return $this->db->insert('analysis_data', $data);
     }
 
-    public function getFileNameModel($file_id)
+    public function getAnalysisHistoryModel($file_id)
     {
-        $this->db->select('fd_name');
-        $this->db->from('file_data');
-        $this->db->where('fd_id', $file_id);
-        return $this->db->get()->row_array();
+        $this->db->select('*');
+        $this->db->from('analysis_data');
+        $this->db->where('ad_fd_id', $file_id);
+        $this->db->limit(10);
+        return $this->db->get()->result_array();
     }
 
     public function updateFileModel($filename, $file_id)
@@ -49,19 +50,38 @@ class AnalysisModel extends CI_Model
         $data = array(
             'fd_name' => $filename,
         );
-        
+
         $this->db->where('fd_id', $file_id);
         return $this->db->update('file_data', $data);
     }
 
     public function deleteResultModel($file_id)
     {
+        if ($this->deleteUploadedFile($file_id) !== false) {
+            $this->db->where('fd_id', $file_id);
+
+            if ($this->db->delete('file_data') !== false) {
+
+                $this->db->where('ad_fd_id', $file_id);
+                return $this->db->delete('analysis_data');
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
+
+    public function deleteUploadedFile($file_id)
+    {
+        $this->db->select('fd_name');
+        $this->db->from('file_data');
         $this->db->where('fd_id', $file_id);
+        $file_name = $this->db->get()->row_array();
 
-        if ($this->db->delete('file_data') !== false) {
-
-            $this->db->where('ad_fd_id', $file_id);
-            return $this->db->delete('analysis_data');
+        if ($file_name !== false || $file_name !== NULL) {
+            unlink('./storage/user-upload/' . $_SESSION['uid'] . '/' . $file_name['fd_name']);
+            return true;
         } else {
             return false;
         }
